Rozdělení modulu spicy a normálních obrázkových příkazů
